requirements.txt
python-telegram-bot==11.1.0
apiai==1.2.3
aldryn-apphooks-config==0.4.1
aldryn-boilerplates==0.7.7
aldryn-categories==1.1.0
aldryn-common==1.0.4
aldryn-disqus==1.0.0
aldryn-events==3.0.0
aldryn-newsblog==2.1.1
aldryn-people==2.1.0
aldryn-translation-tools==0.2.1
Babel==2.6.0
backport-collections==0.1
certifi==2018.10.15
chardet==3.0.4
cmsplugin-contact-plus==1.3.12
cmsplugin-filer==1.1.3
defusedxml==0.5.0
dj-database-url==0.5.0
Django==1.11.16
django-admin-sortable==2.1.9
django-allauth==0.38.0
django-appconf==1.0.2
django-appdata==0.1.6
django-bootstrap3==11.0.0
django-classy-tags==0.8.0
django-cms==3.5.2
django-extended-choices==1.3
django-filer==1.3.2
django-formtools==2.1
django-heroku==0.3.1
django-meta==1.4.1
django-meta-mixin==0.3.0
django-mptt==0.8.7
django-parler==1.9.2
django-phonenumber-field==1.3.0
django-polymorphic==2.0.3
django-recaptcha==1.4.0
django-reversion==1.10.2
django-sekizai==0.10.0
django-simple-math-captcha==1.0.7
django-sortedm2m==1.5.0
django-standard-form==1.1.1
django-tablib==3.2
django-taggit==0.23.0
django-taggit-autosuggest==0.3.3
django-taggit-templatetags==0.2.5
django-templatetag-sugar==1.0
django-treebeard==4.3
django-widget-tweaks==1.4.3
djangocms-admin-style==1.2.8
djangocms-apphook-setup==0.3.0
djangocms-attributes-field==0.3.0
djangocms-column==1.8.0
djangocms-file==2.0.2
djangocms-googlemap==1.1.1
djangocms-installer==1.0.1
djangocms-link==2.1.2
djangocms-picture==2.0.6
djangocms-snippet==2.0.0
djangocms-style==2.0.2
djangocms-text-ckeditor==3.6.0
djangocms-video==2.0.4
easy-thumbnails==2.5
et-xmlfile==1.0.1
future==0.16.0
gunicorn==19.9.0
html5lib==0.9999999
idna==2.7
jdcal==1.4
jsonfield==2.0.2
lxml==4.2.5
oauthlib==2.1.0
odfpy==1.3.6
openpyxl==2.5.9
phonenumbers==8.9.16
phonenumberslite==8.9.16
Pillow==5.3.0
psycopg2==2.7.5
python-dateutil==2.7.3
python-slugify==1.2.0
python3-openid==3.1.0
pytz==2018.5
PyYAML==3.13
requests==2.20.0
requests-oauthlib==1.0.0
reversion==0.2
six==1.11.0
tablib==0.12.1
toml==0.10.0
tzlocal==1.5.1
unicodecsv==0.14.1
Unidecode==0.4.21
urllib3==1.24
whitenoise==4.1
xlrd==1.1.0
xlwt==1.3.0
YURL==0.13
